requirements.txt
fastapi>=0.100.0
uvicorn==0.27.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy>=2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
prometheus-client==0.19.0
plotly
networkx
torch==2.1.2
transformers==4.36.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
spacy==3.7.2
typing-extensions>=4.6.0
tensorflow-macos==2.13.0
pydantic>=2.0.0
mlflow>=2.9.2
fastapi-cache2>=0.2.1
pydantic-settings>=2.0.0